FIX Комшина Дарья. 3822Б1ФИ1. 1_gelu omp

+++ 3822B1FI1/1_gelu_omp/komshina_daria/gelu_omp.cpp
@@ -13,13 +13,13 @@ std::vector<float> GeluOMP(const std::vector<float>& input) {
     const float k1 = 0.7978845608028654f;
     const float k2 = 0.044715f;
 
-    #pragma omp parallel for schedule(static)
+#pragma omp parallel for schedule(static)
     for (std::ptrdiff_t i = 0; i < static_cast<std::ptrdiff_t>(n); ++i) {
         float x = input[static_cast<std::size_t>(i)];
         float x3 = x * x * x;
         float z = k1 * (x + k2 * x3);
 
-        float e = std::expf(2.0f * z);
+        float e = std::exp(2.0f * z);
         float tanh_z = (e - 1.0f) / (e + 1.0f);
 
         out[static_cast<std::size_t>(i)] = k0 * x * (1.0f + tanh_z);
